refactoring(test) : 단위테스트 도메인 로직에 담기도록 수정
