+++ src/main/java/bridge/domain/InputView.java
@@ -10,6 +10,7 @@ public class InputView {
 	private int bridgeLength = 0;
 	private String restartOrQuit = "";
 	private String upOrDown = "";
+
 	/**
 	 * 다리의 길이를 입력받는다.
 	 */
@@ -43,7 +44,7 @@ public class InputView {
 	 * 사용자가 이동할 칸을 입력받는다.
 	 */
 	public String readMoving() {
-		do  {
+		do {
 			System.out.println("\n이동할 칸을 선택해주세요. (위: U, 아래: D)");
 			upOrDown = Console.readLine();
 			try {
@@ -51,7 +52,7 @@ public class InputView {
 			} catch (IllegalArgumentException e) {
 				System.out.println("[ERROR] U와 D만 선택하세요.");
 			}
-		} while(!upOrDown.equals("U") && !upOrDown.equals("D"));
+		} while (!upOrDown.equals("U") && !upOrDown.equals("D"));
 		return upOrDown;
 	}
