+++ src/test/java/bridge/UI_readBridgeSizeTest.java
@@ -1,31 +1,32 @@
-package bridge.domain;
+package bridge;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
-
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-public class UI_readMovingTest extends NsTest {
+import bridge.domain.InputView;
+
+public class UI_readBridgeSizeTest extends NsTest {
 
 	private static final String ERROR_MESSAGE = "[ERROR]";
 	private InputView inputView = new InputView();
 
-	@DisplayName("다리를 선택할 때 U나 D의 문자가 아니면 예외가 발생한다.")
+	@DisplayName("다리의 길이가 3이상 20이하의 정수가 아닐 경우 예외가 발생한다.")
 	@ParameterizedTest
-	@ValueSource(strings = { "F", "q", "R" })
-	public void bridgeShouldBeChoosenUOrD(String UorD) {
+	@ValueSource(strings = { "2", "23" })
+	public void bridgeSizeShouldBeFromThreeToTwentyNumber(String input) {
 		assertSimpleTest(() -> {
-			runException(UorD);
+			runException(input);
 			assertThat(output()).contains(ERROR_MESSAGE);
 		});
 	}
 
 	@Override
 	protected void runMain() {
-		inputView.readMoving();
+		inputView.readBridgeSize();
 	}
 
 }
