+++ src/test/java/bridge/UI_readGameCommandTest.java
@@ -1,30 +1,33 @@
-package bridge.domain;
+package bridge;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
+
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-public class UI_readBridgeSizeTest extends NsTest {
+import bridge.domain.InputView;
+
+class UI_readGameCommandTest extends NsTest {
 
 	private static final String ERROR_MESSAGE = "[ERROR]";
 	private InputView inputView = new InputView();
 
-	@DisplayName("다리의 길이가 3이상 20이하의 정수가 아닐 경우 예외가 발생한다.")
+	@DisplayName("다리 건너기 게임을 재시작할지 종료할지 여부에서 R이나 Q의 문자가 아니면 예외가 발생한다.")
 	@ParameterizedTest
-	@ValueSource(strings = { "2", "23" })
-	public void bridgeSizeShouldBeFromThreeToTwentyNumber(String input) {
+	@ValueSource(strings = { "a", "U" })
+	public void restartShouldBeChoosenROrU(String RorQ) {
 		assertSimpleTest(() -> {
-			runException(input);
+			runException(RorQ);
 			assertThat(output()).contains(ERROR_MESSAGE);
 		});
 	}
 
 	@Override
 	protected void runMain() {
-		inputView.readBridgeSize();
+		inputView.readGameCommand();
 	}
 
 }
