+++ src/test/java/bridge/UI_readMovingTest.java
@@ -1,4 +1,4 @@
-package bridge.domain;
+package bridge;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -8,24 +8,26 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-class UI_readGameCommandTest extends NsTest {
+import bridge.domain.InputView;
+
+public class UI_readMovingTest extends NsTest {
 
 	private static final String ERROR_MESSAGE = "[ERROR]";
 	private InputView inputView = new InputView();
 
-	@DisplayName("다리 건너기 게임을 재시작할지 종료할지 여부에서 R이나 Q의 문자가 아니면 예외가 발생한다.")
+	@DisplayName("다리를 선택할 때 U나 D의 문자가 아니면 예외가 발생한다.")
 	@ParameterizedTest
-	@ValueSource(strings = { "a", "U" })
-	public void restartShouldBeChoosenROrU(String RorQ) {
+	@ValueSource(strings = { "F", "q", "R" })
+	public void bridgeShouldBeChoosenUOrD(String UorD) {
 		assertSimpleTest(() -> {
-			runException(RorQ);
+			runException(UorD);
 			assertThat(output()).contains(ERROR_MESSAGE);
 		});
 	}
 
 	@Override
 	protected void runMain() {
-		inputView.readGameCommand();
+		inputView.readMoving();
 	}
 
 }
