+++ src/test/java/bridge/domain/ApplicationTest.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.domain;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberInRangeTest;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
@@ -6,6 +6,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.util.Lists.newArrayList;
 
 import java.util.List;
+
+import bridge.Application;
 import bridge.domain.BridgeMaker;
 import bridge.domain.BridgeNumberGenerator;
 import camp.nextstep.edu.missionutils.test.NsTest;
