+++ src/test/java/bridge/domain/ApplicationTest2.java
@@ -1,10 +1,12 @@
-package bridge;
+package bridge.domain;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomNumberInRangeTest;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.Test;
+
+import bridge.Application;
 
 public class ApplicationTest2 extends NsTest {
